requirements.txt
certifi==2018.1.18
click==6.7
Flask==0.12.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.2
pandas==0.22.0
PyJWT==1.6.1
python-dateutil==2.7.2
pytz==2018.3
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
Werkzeug==0.14.1
wincertstore==0.2
